Add autoconfiguration for docker models.
