+++ embabel-agent-dependencies/pom.xml
@@ -60,6 +60,12 @@
 
             <dependency>
                 <groupId>com.embabel.agent</groupId>
+                <artifactId>embabel-agent-dockermodels-autoconfigure</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>com.embabel.agent</groupId>
                 <artifactId>embabel-agent-openai-autoconfigure</artifactId>
                 <version>${project.version}</version>
             </dependency>
@@ -127,6 +133,12 @@
             <dependency>
                 <groupId>com.embabel.agent</groupId>
                 <artifactId>embabel-agent-starter-bedrock</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>com.embabel.agent</groupId>
+                <artifactId>embabel-agent-starter-dockermodels</artifactId>
                 <version>${project.version}</version>
             </dependency>
