+++ embabel-agent-starters/embabel-agent-starter-dockermodels/pom.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>com.embabel.agent</groupId>
+        <artifactId>embabel-agent-starters</artifactId>
+        <version>0.1.3-SNAPSHOT</version>
+    </parent>
+    <artifactId>embabel-agent-starter-dockermodels</artifactId>
+    <packaging>jar</packaging>
+    <name>Embabel Agent Docker Models Starter</name>
+    <description>Embabel Agent Docker Models Starter</description>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.embabel.agent</groupId>
+            <artifactId>embabel-agent-platform-autoconfigure</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.embabel.agent</groupId>
+            <artifactId>embabel-agent-dockermodels-autoconfigure</artifactId>
+        </dependency>
+    </dependencies>
+
+</project>
