+++ embabel-agent-starters/pom.xml
@@ -20,6 +20,7 @@
         <module>embabel-agent-starter-openai</module>
         <module>embabel-agent-starter-anthropic</module>
         <module>embabel-agent-starter-ollama</module>
+        <module>embabel-agent-starter-dockermodels</module>
     </modules>
 
 </project>
